refactor: move load step output logic to base action handler
