+++ packages/engine/src/lib/action/action-handler.ts
@@ -1,4 +1,5 @@
-import { StepOutput, Action, ExecutionState, ResumeStepMetadata, ExecutionOutput, ExecutionOutputStatus, StepOutputStatus } from '@activepieces/shared';
+import { StepOutput, Action, ExecutionState, ResumeStepMetadata, ExecutionOutput, ExecutionOutputStatus, StepOutputStatus, StepOutputForActionType } from '@activepieces/shared';
+import { isNil } from 'lodash';
 
 export type ActionHandler = BaseActionHandler
 
@@ -13,6 +14,15 @@ type CtorParams<CA extends Action, RSM extends ResumeStepMetadata = ResumeStepMe
   resumeStepMetadata?: RSM
 }
 
+export type InitStepOutputParams = {
+  executionState: ExecutionState
+}
+
+type LoadStepOutputParams = {
+  executionState: ExecutionState
+  ancestors: [string, number][]
+}
+
 export abstract class BaseActionHandler<CA extends Action = Action, RSM extends ResumeStepMetadata = ResumeStepMetadata> {
   currentAction: CA
   nextAction?: Action
@@ -22,6 +32,36 @@ export abstract class BaseActionHandler<CA extends Action = Action, RSM extends 
     this.currentAction = currentAction
     this.nextAction = nextAction
     this.resumeStepMetadata = resumeStepMetadata
+  }
+
+  /**
+   * initializes an empty step output
+   */
+  protected abstract initStepOutput({ executionState }: InitStepOutputParams): Promise<StepOutputForActionType<CA['type']>>
+
+  /**
+ * Loads old step output if execution is resuming, else initializes an empty step output
+ */
+  protected async loadStepOutput({ executionState, ancestors }: LoadStepOutputParams): Promise<StepOutputForActionType<CA['type']>> {
+    if (isNil(this.resumeStepMetadata)) {
+      return this.initStepOutput({
+        executionState,
+      })
+    }
+
+    const oldStepOutput = executionState.getStepOutput<StepOutputForActionType<CA['type']>>({
+      stepName: this.currentAction.name,
+      ancestors,
+    })
+
+    if (oldStepOutput) {
+      oldStepOutput.status = StepOutputStatus.RUNNING
+      delete oldStepOutput.pauseMetadata
+    }
+
+    return oldStepOutput ?? this.initStepOutput({
+      executionState,
+    })
   }
 
   protected handleFlowExecutorOutput({ executionOutput, stepOutput }: HandleFlowExecutorOutput) {
